requirements.txt
pandas
python-telegram-bot
APScheduler
openpyxl
fpdf
